requirements.txt
python-dotenv>=0.19.0
Flask>=2.0.1
flask-cors>=3.0.10
requests>=2.25.1
